vwaitforio(9): tsleep(9) -> tsleep_nsec(9); ok visa@

+++ sys/sys/vnode.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: vnode.h,v 1.150 2019/02/17 22:17:28 tedu Exp $	*/
+/*	$OpenBSD: vnode.h,v 1.151 2019/07/19 00:54:59 cheloha Exp $	*/
 /*	$NetBSD: vnode.h,v 1.38 1996/02/29 20:59:05 cgd Exp $	*/
 
 /*
@@ -600,7 +600,7 @@ void	vgonel(struct vnode *, struct proc *);
 int	vinvalbuf(struct vnode *, int, struct ucred *, struct proc *,
 	    int, int);
 void	vntblinit(void);
-int	vwaitforio(struct vnode *, int, char *, int);
+int	vwaitforio(struct vnode *, int, char *, uint64_t);
 void	vwakeup(struct vnode *);
 void	vput(struct vnode *);
 int	vrecycle(struct vnode *, struct proc *);
